Creación carpetas
Se realizan cambios en el main y se agregan las carpetas para hacer el codigo de bajo consumo y de interrupciones

+++ Lab1/Bajo consumo/Practica_1.X/displays.c
@@ -223,7 +223,7 @@ void Ds_Convertir_en_uni(Ds_Display *dsp, uint16_t voltaje){
 
 uint16_t Ds_Conversor_ADC (uint16_t adcOUT){
     float pendiente = 0.97665;
-    uint16_t voltios = (uint16_t) adcOUT*pendiente;
+    uint16_t voltios = (uint16_t)(((float)adcOUT)*pendiente);
 	return(voltios);
 }
 
@@ -238,7 +238,6 @@ void Ds_Procese_displays (Ds_Display *dsp){
                 dsp->estados =D1A;
                 dsp->tempA=dsp->tiempoApagado;
             }
-            
 
             Ds_Encienda_Disp(1);
             Ds_BCD(dsp->D1);
@@ -250,8 +249,8 @@ void Ds_Procese_displays (Ds_Display *dsp){
             if(!(dsp->tempA)){
                 dsp->estados =D2E;
                 dsp->tempE=dsp->tiempoEncendido;
-            }
-            
+            } 
+
             Ds_Encienda_Disp(0);
         break;
 
@@ -261,7 +260,7 @@ void Ds_Procese_displays (Ds_Display *dsp){
                 dsp->estados =D2A;
                 dsp->tempA=dsp->tiempoApagado;
             }
-        
+
             Ds_Encienda_Disp(2); 
             Ds_BCD(dsp->D2);
         break;
@@ -300,8 +299,8 @@ void Ds_Procese_displays (Ds_Display *dsp){
         default:
             while(1){
 
-            Ds_BCD(10);
-			Ds_Encienda_Disp(2);
+                Ds_BCD(10);
+			    Ds_Encienda_Disp(2);
             }
         }
     }
